Update source GLFW revision in README.
It might be good time to start making a development script that updates
the glfw/ subfolder and writes the GLFW revision into some file, so
that it's not possible to forgot to update it.
Not to mention it was relatively hard to find the issue that described
steps to update. For reference, it was:
https://github.com/go-gl/glfw3/issues/82#issuecomment-53175925