README: Update installation steps, mention dependencies, remove older comments.
Use -u flag for go get command to ensure latest version is installed (latest versions tend to exist in order to fix bugs and issues).

Mention that dependencies of GLFW (C library) need to be installed (but not GLFW itself, that is built as part of Go package).

Remove old comments like "we're breaking the API" as that's a constant now; the changelog remains below. Also remove comment about adding examples since the README in examples repo does a better job of communicating how to best contribute there. Change title from "bindings" since it's now a more complete Go package, not just bindings.

Use a godoc badge for the documentation link.

Fixes #138 (and related to #140).