v3.2/glfw: Use and document *image.NRGBA as expected image format.
GLFW images are non-premultiplied RGBA, which corresponds to
*image.NRGBA type. This was previously not documented in GLFW,
but has since been fixed in glfw/glfw#426 and glfw/glfw#1027.

This is not a breaking API change and functionality is preserved.
However, for optimal performance (to avoid unnecessary conversions),
images should now ideally be provided as *image.NRGBA rather than
*image.RGBA.

Fixes #209.