Add scripts/grab-upstream.sh
The purpose of this script is to make it self evident from Travis' status that
the files in the repository match upstream.

It also means that if you want to update the code, you just need to change
GLFW_C_REVISION.txt and run scripts/grab-upstream.sh.

Since it is checked, it will require less review to make such changes in the future.

Since it is automated, it will be easy for someone to update it if they need to
and not make mistakes regarding which files are included or not.