.travis.yml: remove default travis_install_go_dependencies install step
We have a custom implementation that we want to run instead of the
default install step, so just run "true" which exits with 0 exit code.

Having a command in addition to a comment wasn't needed to disable the
default install step in the past, but for some reason, it is needed now.

Reference: https://docs.travis-ci.com/user/languages/go/#dependency-management