Make build cache a function of the glfw C sources
This is a workaround for the issue described in #269. Go unfortunately doesn't
consider the C source files as part of its build cache calculation. Therefore,
invalidate it by writing a hash which is computed across the full C source
directory.

This hash is only updated when `go generate` is called, but this will at least
help the majority of users, when glfw is updated. This is encoded in
grab-upstream.sh which is used for updating the sources.

This hash is intended to be protected from divergance by CI, too.

Resolves #269.