Add FreeBSD support to the glfw joystick code.
Tested with the multimedia/webcamd port as "driver", and an
Xbox One game pad. Assumes "Linux" for game pad definitions,
which seems right as webcamd actually is wrapping the Linux
drivers.

Note that part of this may have to be synced upstream with glfw.
No idea how this process works in go-gl/glfw.