v3.3/glfw: update to GLFW 3.3.4
GLFW 3.3.4 was released on April 8, 2021.
See https://github.com/glfw/glfw/releases/tag/3.3.4.

Done with scripts/grab-upstream.sh (albeit with edits):

	echo 'b3e0aae393ef6c5cda7dcad0cba06bef23a1dda9' > v3.3/glfw/GLFW_C_REVISION.txt  # 3.3.4.
	./scripts/grab-upstream.sh

The grab-upstream.sh script was modified to comment out Wayland
regeneration and invocations of sed, as it didn't work on macOS
out of the box. (I didn't think it'd be worth it given the diff
of 3.3.3 → 3.3.4 being so small.)

Fixes #328.