v3.3/glfw: remove 'immediately' from internal error message
An internal error is an error that should never happen, but it can still
happen when there is a bug in the code. When it happens, glfw prints a
suggestion of reporting the bug, as that way we can learn about the bug
and fix it.

The word "immediately" may create a sense of urgency, but in fact these
bug reports are not very time sensitive, at least no more than any other
issue. Relax the language a bit.

Also, as commonly done, print these errors to stderr rather than stdout.